sample_indices: add new seq::index module for index sampling
